Hakyll.Web.Dhall

Description

Hakyll compiler and loader for Dhall files. Functions are intended to track all local dependencies within the project directory, so rebuilds are properly triggered on up-stream imports. Provides options for customizing rebuilding behavior for network, environment variable, and non-project local files.

Configuration and Options

data DhallCompilerOptions a

Options for loading Dhall files

Constructors

DCO

Fields

dcoResolver :: DhallResolver a

Method to resolve imports encountered in files. See documentation of DhallResolver for more details.

dcoTrust :: Set DhallCompilerTrust

Set of "trusted" import behaviors. Files with external references or imports that aren't described in this set are always rebuilt every time.

Default: singleton DCTRemote

That is, do not trust any dependencies on the local disk outside of the project directory, but trust that any URL imports remain unchanged.

dcoMinimize :: Bool

Strictly for usage with dhallCompiler and family: should the result be "minimized" (all in one line) or pretty-printed for human readability?

Can be useful for saving bandwidth.

Default: False

dcoNormalize :: Bool

If True, reduce expressions to normal form before using them.

Default: True

data DhallCompilerTrust

Types of external imports that a Dhall file may have.

Constructors

DCTLocal
File on local filesystem outside of project directory, and therefore not tracked by Hakyll

DCTRemote
Link to remote resource over a network connection

DCTEnv
Reference to environment variable on machine

defaultDhallCompilerOptions :: DefaultDhallResolver a => DhallCompilerOptions a

Default DhallCompilerOptions. If the type variable is not inferrable, it can be helpful to use TypeApplications syntax:

defaultCompilerOptions @Import         -- do not resolve imports
defaultCompilerOptions @X              -- resolve imports

Resolver Behaviors

data DhallResolver :: Type -> Type where

Method for resolving imports.

The choice will determine the type of expression that loadDhallExpr and family will produce.

Constructors

DRRaw :: {..} -> DhallResolver Import
Leave imports as imports, but optionally remap the destinations.

Default: leave imports unchanged

Fields

drRemap :: Import -> Compiler Import

DRFull :: DhallResolver X
Completely resolve all imports in IO

class DefaultDhallResolver a where

Helper typeclass to allow functions to be polymorphic over different DhallResolver types.

Provides default behavior for each resolver type.

Minimal complete definition

defaultDhallResolver

Methods

defaultDhallResolver :: DhallResolver a

Load Dhall Files

As as custom Haskell types

loadDhall :: Type a -> Identifier -> Compiler (Item a)

Load a value of type a that is parsed from a Dhall file at the given Identifier.

loadDhallWith :: DhallCompilerOptions X -> Type a -> Identifier -> Compiler (Item a)

Version of loadDhall taking custom DhallCompilerOptions.

As raw expressions

loadDhallExpr :: DefaultDhallResolver a => Identifier -> Compiler (Item (Expr Src a))

Load and parse the body of the given Identifier as a Dhall expression.

If you wrap the result in DExpr, you can save the result as a snapshot.

loadDhallExprWith :: DhallCompilerOptions a -> Identifier -> Compiler (Item (Expr Src a))

Version of loadDhallExpr taking custom DhallCompilerOptions.

newtype DExpr a

Newtype wrapper over Expr Src a (A Dhall expression) with an appropriate Binary instance, meant to be usable as a compilable Hakyll result that can be saved with saveSnapshot, load, etc.

Constructors

DExpr

Fields

getDExpr :: Expr Src a

Compile Dhall Files

dhallCompiler :: forall a. (DefaultDhallResolver a, Pretty a) => Compiler (Item String)

Compile the Dhall file as text according to default DhallCompilerOptions. Note that this is polymorphic over both "raw" and "fully resolved" versions; it must be called with TypeApplications

dhallRawCompiler  = dhallCompiler Import
dhallFullCompiler = dhallCompiler X

It might be more convenient to just use dhallRawCompiler or dhallFullCompiler.

dhallRawCompiler :: Compiler (Item String)

Compile the Dhall file as text according to default DhallCompilerOptions while leaving all imports unchanged and unresolved.

dhallFullCompiler :: Compiler (Item String)

Compile the Dhall file as text according to default DhallCompilerOptions, resolving all imports in IO.

dhallCompilerWith :: Pretty a => DhallCompilerOptions a -> Compiler (Item String)

dhallCompiler, but with custom DhallCompilerOptions.

Internal Utilities

parseDhallWith :: DhallCompilerOptions a -> Identifier -> Text -> Compiler (Item (Expr Src a))

Parse a Dhall source.

This encapsulates the "magic" of tracking dependencies. Any local dependencies within the project directory are tracked by Hakyll, and so modifications to required files will also cause upstream files to be rebuilt.

parseRawDhallWith :: DhallCompilerOptions Import -> Identifier -> Text -> Compiler (Item (Expr Src Import))

Version of parseDhallWith that only acceps the DRRaw resolver, remapping the imports with the function in the DRRaw.

resolveDhallImports :: DhallCompilerOptions X -> Identifier -> Expr Src Import -> Compiler (Expr Src X)

Resolve all imports in a parsed Dhall expression.

Implemented so that any local dependencies within the project directory are tracked by Hakyll, and so modifications to required files will also cause upstream files to be rebuilt.